move MarkDownSettingContents to  admin dir
