+++ src/components/Admin/Markdown/MarkDownSettingContents.tsx
@@ -3,9 +3,9 @@ import { Card, CardBody } from 'reactstrap';
 
 import { useTranslation } from '~/i18n';
 
-import LineBreakForm from './LineBreakForm';
-import PresentationForm from './PresentationForm';
-import XssForm from './XssForm';
+import LineBreakForm from '../../../client/js/components/Admin/MarkdownSetting/LineBreakForm';
+import PresentationForm from '../../../client/js/components/Admin/MarkdownSetting/PresentationForm';
+import XssForm from '../../../client/js/components/Admin/MarkdownSetting/XssForm';
 
 type Props = {
 };
